// web/src/api/generated/model/sseFinalStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Mixd
 * Personal music metadata hub
 * OpenAPI spec version: 0.11.7
 */

export type SseFinalStatus = typeof SseFinalStatus[keyof typeof SseFinalStatus];


export const SseFinalStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
  crashed: 'crashed',
} as const;