// web/src/api/generated/model/sseSubOperationOutcome.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Mixd
 * Personal music metadata hub
 * OpenAPI spec version: 0.11.7
 */

export type SseSubOperationOutcome = typeof SseSubOperationOutcome[keyof typeof SseSubOperationOutcome];


export const SseSubOperationOutcome = {
  succeeded: 'succeeded',
  skipped_unchanged: 'skipped_unchanged',
  failed: 'failed',
} as const;